requirements.txt
pandas
snowflake-connector-python
matplotlib
snowflake-connector-python[pandas]
pyarrow